Beta 1.2.1
Index add filelist
