+++ article/wp-content/themes/twentyseventeen/header.php
@@ -15,7 +15,7 @@
 ?>
 <?php
 require_once("../../conn/conn.php");
-require_once("../tools/history.php");
+require_once("../tools/request_history.php");
 ?><!-- 保存访问信息，可复用 -->
 <!DOCTYPE html>
 <html <?php language_attributes(); ?> class="no-js no-svg">
